Generic Register Driver
Generic Register Driver
Summary:
The goal of this Assignment is to test how to build a generic service class that can be used to provide an abstraction for the hardware registers
Problem Statement:
We have different microcontrollers, each of them has different register size and different register permissions based on the processor architecture.
Our goal is to implement a generic class to abstract the user.
Requirements:
Class APIs:
setRegisterValue ⇒ Sets the whole register bits with the value sent by the user
getRegisterValue ⇒ Reads the whole register value
setBit ⇒ set only a specific bit in the register by 1
clearBit ⇒ set only a specific bit in the register by 0
getBit ⇒ get only the value of a specific bit in the register
setPermission ⇒ change the register permission of certain bit of the register or the whole register (this API can be used any time during the application running)
The register permission could be:
Read only
Write only
Read and Write
Reserved
This class should be able to deal with any register size
Your class should handle any user violation or misuse and return a suitable error status